Pretty good YNAB import population for transactions. Still few todos left.
ToDos:
- Investigate and implement hierarchical category population (right ID referencing)
- Investigate and implement split transactions

- Budget is completely missing

- And of course there may be cases not considered, because sample data only reflects my budgeting scenarios
==> Thus, all ways to budget possible with YNAB have to be considered. (But this has a rather low priority at this moment)

+++ BFF/Model/Native/Budget.cs
@@ -17,7 +17,7 @@ namespace BFF.Model.Native
         public override string CreateTableStatement { get; }
 
         [Key]
-        public override int ID { get; set; }
+        public override long ID { get; set; }
         
         public DateTime MonthYear { get; set; }
